+++ Passepartout/App/Views/OrganizerView.swift
@@ -27,6 +27,17 @@ import SwiftUI
 import PassepartoutLibrary
 
 struct OrganizerView: View {
+    enum ModalType: Identifiable {
+        case interactiveAccount(profile: Profile)
+
+        // XXX: alert ids
+        var id: Int {
+            switch self {
+            case .interactiveAccount: return 1
+            }
+        }
+    }
+
     enum AlertType: Identifiable {
         case subscribeReddit
         
@@ -44,6 +55,8 @@ struct OrganizerView: View {
 
     @State private var addProfileModalType: AddProfileMenu.ModalType?
 
+    @State private var modalType: ModalType?
+
     @State private var alertType: AlertType?
 
     @State private var isHostFileImporterPresented = false
@@ -58,7 +71,7 @@ struct OrganizerView: View {
         debugChanges()
         return ZStack {
             hiddenSceneView
-            ProfilesList()
+            ProfilesList(modalType: $modalType)
         }.toolbar {
             ToolbarItem(placement: .primaryAction) {
                 AddProfileMenu(
@@ -71,7 +84,8 @@ struct OrganizerView: View {
                     SettingsButton()
                 }
             }
-        }.alert(item: $alertType, content: presentedAlert)
+        }.sheet(item: $modalType, content: presentedModal)
+        .alert(item: $alertType, content: presentedAlert)
         .fileImporter(
             isPresented: $isHostFileImporterPresented,
             allowedContentTypes: hostFileTypes,
@@ -117,6 +131,16 @@ extension OrganizerView {
     
     private func onOpenURL(_ url: URL) {
         addProfileModalType = .addHost(url, false)
+    }
+
+    @ViewBuilder
+    private func presentedModal(_ modalType: ModalType) -> some View {
+        switch modalType {
+        case .interactiveAccount(let profile):
+            NavigationView {
+                InteractiveConnectionView(profile: profile)
+            }.themeGlobal()
+        }
     }
 
     private func presentedAlert(_ alertType: AlertType) -> Alert {