+++ Passepartout/AppShared/L10n/Unlocalized.swift
@@ -252,5 +252,7 @@ enum Unlocalized {
 
     enum Other {
         static let siri = "Siri"
+        
+        static let totp = "TOTP"
     }
 }